requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
adal=1.2.7=pypi_0
aiofiles=23.1.0=pypi_0
aiohttp=3.8.3=pypi_0
aiokafka=0.8.0=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.9.3=pypi_0
altair=4.2.2=pypi_0
anyio=3.6.2=pypi_0
argcomplete=2.0.0=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
azure-common=1.1.28=pypi_0
azure-core=1.26.3=pypi_0
azure-graphrbac=0.61.1=pypi_0
azure-mgmt-authorization=3.0.0=pypi_0
azure-mgmt-containerregistry=10.0.0=pypi_0
azure-mgmt-core=1.3.2=pypi_0
azure-mgmt-keyvault=10.1.0=pypi_0
azure-mgmt-resource=21.2.1=pypi_0
azure-mgmt-storage=20.1.0=pypi_0
azureml-core=1.48.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports-tempfile=1.0=pypi_0
backports-weakref=1.0.post1=pypi_0
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bcrypt=4.0.1=pypi_0
blas=1.0=mkl
blinker=1.5=pypi_0
boto3=1.26.68=pypi_0
botocore=1.29.68=pypi_0
brotli=1.0.9=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.7=pypi_0
cryptography=38.0.4=py310h9ce1e76_0
cycler=0.11.0=pypi_0
databricks-cli=0.17.4=pypi_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
distlib=0.3.6=pypi_0
docker=6.0.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fastapi=0.88.0=pypi_0
filelock=3.9.0=pypi_0
flask=2.2.2=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
future=0.18.2=py310h06a4308_1
gevent=22.10.2=pypi_0
geventhttpclient=2.0.2=pypi_0
giflib=5.2.1=h5eee18b_1
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
google-api-core=2.11.0=pypi_0
google-auth=2.16.0=pypi_0
google-cloud-core=2.3.2=pypi_0
google-cloud-storage=2.7.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.4.1=pypi_0
googleapis-common-protos=1.58.0=pypi_0
greenlet=2.0.2=pypi_0
grpcio=1.51.1=pypi_0
gunicorn=20.1.0=pypi_0
h11=0.14.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=py310h06a4308_0
importlib-metadata=5.2.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.9.0=pyh41d4057_0
isodate=0.6.1=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
jeepney=0.8.0=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
jsonpickle=2.2.0=pypi_0
jsonschema=4.17.3=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.2.0=py310hff52083_0
kafka-python=2.0.2=pypi_0
kiwisolver=1.4.4=pypi_0
knack=0.10.1=pypi_0
kubernetes=25.3.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_1
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
llvmlite=0.39.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
mako=1.2.4=pypi_0
markdown=3.4.1=pypi_0
markdown-it-py=2.1.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mlflow=2.1.1=pypi_0
mlserver=1.3.0.dev2=pypi_0
mlserver-mlflow=1.3.0.dev2=pypi_0
msal=1.21.0=pypi_0
msal-extensions=1.0.0=pypi_0
msrest=0.7.1=pypi_0
msrestazure=0.6.4=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
ndg-httpsclient=0.5.1=pypi_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numba=0.56.4=pypi_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
oauthlib=3.2.2=pypi_0
openssl=1.1.1s=h7f8727e_0
orjson=3.8.6=pypi_0
packaging=21.3=pypi_0
pandas=1.5.3=pypi_0
paramiko=2.12.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.11.0=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=pypi_0
pip=22.3.1=py310h06a4308_0
pkginfo=1.9.6=pypi_0
platformdirs=3.0.0=pyhd8ed1ab_0
portalocker=2.7.0=pypi_0
prometheus-client=0.16.0=pypi_0
prometheus-flask-exporter=0.21.0=pypi_0
prompt-toolkit=3.0.36=pyha770c72_0
protobuf=3.20.3=pypi_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-grpc-prometheus=0.7.0=pypi_0
pyarrow=10.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.4=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.14.0=pyhd8ed1ab_0
pyjwt=2.6.0=pypi_0
pympler=1.0.1=pypi_0
pymysql=1.0.2=pypi_0
pynacl=1.5.0=pypi_0
pyopenssl=22.1.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysftp=0.2.9=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.9=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.21.1=pypi_0
python-rapidjson=1.9=pypi_0
python_abi=3.10=2_cp310
pytorch=1.12.1=cpu_py310hb1f1ab4_1
pytz=2022.7.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=py310h5eee18b_1
pyzmq=23.2.0=py310h6a678d5_0
querystring-parser=1.2.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
requests-auth-aws-sigv4=0.7=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.3.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pypi_0
scikit-learn=1.2.1=pypi_0
scipy=1.10.0=pypi_0
secretstorage=3.3.3=pypi_0
semver=2.13.0=pypi_0
setuptools=65.6.3=py310h06a4308_0
shap=0.41.0=pypi_0
six=1.16.0=pyh6c4a22f_0
slicer=0.0.7=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
sqlalchemy=1.4.46=pypi_0
sqlite=3.40.1=h5082296_0
sqlparse=0.4.3=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.22.0=pypi_0
starlette-exporter=0.15.1=pypi_0
streamlit=1.17.0=pypi_0
tabulate=0.9.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchvision=0.13.1=cpu_py310h164cc8f_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
tritonclient=2.30.0=pypi_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022.7=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.14=py310h06a4308_0
uvicorn=0.20.0=pypi_0
uvloop=0.17.0=pypi_0
validators=0.20.0=pypi_0
virtualenv=20.19.0=pypi_0
watchdog=2.2.1=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.8.2=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.12.1=pypi_0
zlib=1.2.13=h5eee18b_0
zope-event=4.6=pypi_0
zope-interface=5.5.2=pypi_0
zstd=1.5.2=ha4553b6_0
